+++ src/main/java/dev/smootheez/elytracontrol/config/ElytraControlConfig.java
@@ -1,18 +1,20 @@
 package dev.smootheez.elytracontrol.config;
 
-import dev.smootheez.elytracontrol.Constants;
-import dev.smootheez.elytracontrol.config.option.LockIconMode;
-import dev.smootheez.scl.annotation.Config;
-import dev.smootheez.scl.api.ConfigProvider;
-import dev.smootheez.scl.config.ConfigOption;
+import dev.smootheez.elytracontrol.*;
+import dev.smootheez.elytracontrol.config.option.*;
+import dev.smootheez.scl.api.*;
+import dev.smootheez.scl.config.*;
 
-@Config(Constants.MOD_ID)
-public class ElytraControlConfig implements ConfigProvider {
-    public static final ConfigOption<Boolean> elytraLock = ConfigOption.create("elytraLock", true);
-    public static final ConfigOption<Boolean> elytraCancel = ConfigOption.create("elytraCancel", true);
-    public static final ConfigOption<Boolean> easyFlight = ConfigOption.create("easyFlight", false);
-    public static final ConfigOption<Boolean> easyFlightMessage = ConfigOption.create("easyFlightMessage", true);
-    public static final ConfigOption<Boolean> elytraLockMessage = ConfigOption.create("elytraLockMessage", true);
+@Config(name = Constants.MOD_ID, gui = true)
+public class ElytraControlConfig {
+    public static final ConfigOption<Boolean> DEFAULT_ELTRA_CONTROL = ConfigOption.create("defaultElytraControl", true);
 
-    public static final ConfigOption<LockIconMode> lockIconMode = ConfigOption.create("iconLockMode", LockIconMode.ICON_TEXT);
+    public static final ConfigOption<Boolean> ALLOW_FLYING = ConfigOption.create("allowFlying", true);
+    public static final ConfigOption<Boolean> DISABLE_FLY_NOTIFICATION = ConfigOption.create("disableFlyNotification", true);
+
+    public static final ConfigOption<Boolean> EASY_FLY = ConfigOption.create("easyFly", false);
+    public static final ConfigOption<Boolean> EASY_FLY_NOTIFICATION = ConfigOption.create("easyFlyNotification", true);
+
+    public static final ConfigOption<LockIconMode> LOCK_ICON_MODE = ConfigOption.create("lockIconMode", LockIconMode.ICON_TEXT);
+    public static final ConfigOption<OverlayPosition> OVERLAY_POSITION = ConfigOption.create("overlayPosition", OverlayPosition.TOP_LEFT);
 }