+++ src/main/java/dev/smootheez/elytracontrol/helpers/PlayerEntityHelper.java
@@ -1,9 +1,0 @@
-package dev.smootheez.elytracontrol.helpers;
-
-import net.minecraft.entity.player.PlayerEntity;
-
-public class PlayerEntityHelper {
-    public static boolean isPlayerGliding(PlayerEntity player) {
-        return player.isGliding();
-    }
-}