+++ src/main/java/dev/smootheez/elytracontrol/util/ActionStep.java
@@ -1,0 +1,7 @@
+package dev.smootheez.elytracontrol.util;
+
+public enum ActionStep {
+    START_MOMENTUM,
+    START_FLYING,
+    USE_FIREWORK
+}